feat!: add structured logger
The current logger is not flexible enough to expose more information on
the operation. Now the users can provide any kind of logger they have as
long as there is a driver for it in logr.

Users using the Logger function should use the DeprecatedLogger
functions instead. The DeprecatedLogger will arrive in the next commit!